+++ packages/core/src/index.ts
@@ -3,9 +3,12 @@
  */
 
 export * from "./types";
+export { FRAMEWORK_PRESETS, applyFramework } from "./framework";
+export type { FrameworkPreset } from "./framework";
 export { parseJson, JsonParseError } from "./json/parse";
 export type { ParsedJson } from "./json/parse";
 export { flatten } from "./flatten";
+export type { FlattenOptions } from "./flatten";
 export { compareKeys } from "./key-comparator/compare";
 export type { KeyDiff } from "./key-comparator/compare";
 export {